we no longer need the redundant hybrid/api.py
