+++ src/api/api_codegen.cc
@@ -9,8 +9,6 @@
 #include <tvm/lowered_func.h>
 #include <tvm/api_registry.h>
 
-#include "../codegen/codegen_hybrid.h"
-
 namespace tvm {
 namespace codegen {
 
@@ -26,19 +24,6 @@ TVM_REGISTER_API("codegen._Build")
 TVM_REGISTER_API("module._PackImportsToC")
 .set_body([](TVMArgs args, TVMRetValue *ret) {
     *ret = PackImportsToC(args[0], args[1]);
-  });
-
-TVM_REGISTER_API("codegen._HybridDump")
-.set_body([](TVMArgs args, TVMRetValue *ret) {
-    Stmt stmt;
-    if (args[0].IsNodeType<Stmt>()) {
-      stmt = args[0];
-    } else if (args[0].IsNodeType<Expr>()) {
-      stmt = Evaluate::make(args[0]);
-    }
-    CodeGenHybrid generator;
-    generator.PrintStmt(stmt);
-    *ret = generator.Finish();
   });
 }  // namespace codegen
 }  // namespace tvm
